Добавляет api.js
