+++ js/form.js
@@ -2,25 +2,12 @@ import { isEscapeKey } from './util.js';
 import { changeEffect, resetFilter, createSlider } from './effects.js';
 import { addValidator, resetPristine, validatePristine } from './validate.js';
 import { activateScale, resetScale } from './scaling.js';
-// import { sendData } from './data.js';
-// import { renderFailMessage, renderSuccessMessage } from './messages.js';
-
-// const GET_URL = 'https://29.javascript.pages.academy/kekstagram/data';
 
 const form = document.querySelector('.img-upload__form');
 const overlay = document.querySelector('.img-upload__overlay');
 const buttonCancel = document.querySelector('#upload-cancel');
 const fileField = document.querySelector('#upload-file');
 const effectsField = document.querySelector('.effects');
-
-// const onSendSuccess = () => {
-//   renderSuccessMessage();
-//   closeForm();
-// };
-
-// const onSendFail = () => {
-//   renderFailMessage();
-// };
 
 const onDocumentKeydown = (event) => {
   if (isEscapeKey(event) && !event.target.closest('.text__hashtags') &&
@@ -39,13 +26,6 @@ const openForm = () => {
 const onButtonCancelClick = () => closeForm();
 const onFileInputChange = () => openForm();
 const onEffectsChange = (event) => changeEffect(event);
-
-// const onFormSubmit = (event) => {
-//   event.preventDefault();
-//   if (validatePristine()) {
-//     sendData(GET_URL, onSendSuccess, onSendFail, new FormData(event.target));
-//   }
-// };
 
 const onFormSubmit = (event) => {
   if (!validatePristine()) {
