+++ js/main.js
@@ -1,5 +1,6 @@
 import { renderPosts } from './render-posts.js';
 import { setFormAction } from './form.js';
+import { getData } from './api.js';
 
 
 renderPosts();
