--- js/message.js
@@ -1,55 +1,0 @@
-import { isEscapeKey } from './util.js';
-
-const failMessageTemplate = document.querySelector('#error').content.querySelector('.error');
-const successMessageTemplate = document.querySelector('#success').content.querySelector('.success');
-let failMessageClone;
-let successMessageClone;
-
-const closeFailMessage = () => {
-  failMessageClone.remove();
-  document.removeEventListener('keydown', onDocumentKeydown);
-  failMessageClone = '';
-};
-
-const closeSuccessMessage = () => {
-  successMessageClone.remove();
-  document.removeEventListener('keydown', onDocumentKeydown);
-  successMessageClone = '';
-};
-
-function onDocumentKeydown(event) {
-  if(isEscapeKey(event)) {
-    event.preventDefault();
-    if (failMessageClone) {
-      closeFailMessage();
-      return;
-    }
-    closeSuccessMessage();
-  }
-}
-
-const onFailMessageButtonClick = (evt) => {
-  evt.preventDefault();
-  closeFailMessage();
-};
-
-const onSuccessMessageButtonClick = (evt) => {
-  evt.preventDefault();
-  closeSuccessMessage();
-};
-
-const renderFailMessage = () => {
-  failMessageClone = failMessageTemplate.cloneNode(true);
-  document.body.append(failMessageClone);
-  document.addEventListener('keydown', onDocumentKeydown);
-  failMessageClone.querySelector('.error__button').addEventListener('click', onFailMessageButtonClick);
-};
-
-const renderSuccessMessage = () => {
-  successMessageClone = successMessageTemplate.cloneNode(true);
-  document.body.append(successMessageClone);
-  document.addEventListener('keydown', onDocumentKeydown);
-  successMessageClone.querySelector('.success__button').addEventListener('click', onSuccessMessageButtonClick);
-};
-
-export {renderFailMessage, renderSuccessMessage};
